+++ src/components/Input/NumberInput/NumberInput.types.ts
@@ -1,19 +1,18 @@
 import type { InputProps } from "../Input.types";
-import type { Nullable } from "@ubloimmo/front-util";
 
 export type NumberInputProps = InputProps<"number"> & {
   /**
    * The input's minimum value.
    *
-   * @default undefined
+   * @default -Infinity
    */
-  min?: Nullable<number>;
+  min?: number;
   /**
    * The input's maximum value.
    *
-   * @default undefined
+   * @default Infinity
    */
-  max?: Nullable<number>;
+  max?: number;
   /**
    * The input step size.
    * Affects the value increment/decrement when using the arrow keys and controls
@@ -23,6 +22,12 @@ export type NumberInputProps = InputProps<"number"> & {
    * @default 1
    */
   step?: number;
+  /**
+   * The number's scale.
+   * Will multiply the number's precision by 10^scale
+   * @default 0
+   */
+  scale?: number;
 };
 
 export type DefaultNumberInputProps = Required<NumberInputProps>;